seek fixes
NUMERIC OR IS JUST ONE PIPE!

+++ ArtOS_lib/kernel.h
@@ -59,7 +59,7 @@ typedef struct event_t event_t;
 // files
 int write(int fd, const char* buf, unsigned long count);
 int read(int fd, char* buf, size_t count);
-int seek(int fd, i64 offset, int whence);
+i64 seek(int fd, i64 offset, int whence);
 int open(const char* pathname, int flags);
 int close(const int fd);
 void _exit(int status);
